Analytics and classField error
Added Analytics initialization to be tested later. Also made the Name
textfield not as annoying when there is nothing entered into it.
Removed the alert and added a shake effect.

Fix shake effect + analytic events

Shake effect was shaking the entire inputs, updated constraints to stop
that. Attempted to add analytic event when adding a class to capture
the class data, not entirely sure if the feature works yet.

Updated Analytics

Events are firing and confirmed to be reporting to the Flurry server.
Events are only recorded when the app is sent into the background
first, before stopping via Xcode.